PropertyDecorator: Recover per-language values after missing parents
If a file for a node is missing a parent in the same language, the
values from the default language got injected, which switched
languages for subnodes.

For example if the property `prop1` has value `foo` globally, and the
root node has a `fr.md` that sets `prop1` to `bar`. Then if the
default language is `fr`, any `en.md` for a node whose the parent node
is missing the `en.md` but has a `fr.md` sees `prop1` as `bar`
(instead of `foo`).

So instead of inheriting all variables that a parent node has, we only
inherit those that the parent node explicitly sets. This allows to
more seamlessly recover from missing parents.